+++ mail/mail-common/src/user_context.rs
@@ -4,6 +4,7 @@ mod events;
 mod images;
 mod initialization;
 
+use crate::actions::draft::SEND_ACTION_GROUP;
 use crate::models::{Conversation, Message};
 use crate::prefetch::{Prefetch, PrefetchNotify};
 use crate::user_context::action_queue::new_action_queue;
@@ -11,7 +12,7 @@ use crate::user_context::cache::{Cache, CacheAttachmentConfig, CacheMessageConfi
 use crate::{AppError, MailContext, MailContextError, MailContextResult};
 use anyhow::anyhow;
 pub use initialization::*;
-use proton_action_queue::queue::{Queue, QueueExecutor, QueuedResult};
+use proton_action_queue::queue::{Queue, QueueAutoExecutor, QueueAutoExecutorPool};
 use proton_api_core::auth::UserKeySecret;
 use proton_api_core::connection_status::ConnectionStatus;
 use proton_api_core::crypto_clock;
@@ -31,6 +32,7 @@ use proton_event_loop::foreground_loop::EventLoop;
 use stash::orm::Model;
 use stash::stash::{Bond, Stash, Tether};
 use std::future::Future;
+use std::num::NonZeroUsize;
 use std::path::PathBuf;
 use std::sync::{Arc, OnceLock, Weak};
 use std::time::Duration;
@@ -45,7 +47,8 @@ pub struct MailUserContext {
     cache: Cache,
     event_loop: EventLoop,
     action_queue: Queue,
-    default_queue_executor: QueueExecutor,
+    default_queue_executor: QueueAutoExecutor,
+    send_queue_executors: QueueAutoExecutorPool,
     prefetch: PrefetchNotify,
 }
 
@@ -60,6 +63,11 @@ impl MailUserContext {
         let cache = Cache::new(cache_path, mail_context.mail_cache_size).await?;
         let action_queue = new_action_queue(stash).await?;
         let queue_executor = action_queue.new_executor();
+        let send_queue_executors = QueueAutoExecutorPool::new(
+            &action_queue,
+            &SEND_ACTION_GROUP,
+            NonZeroUsize::new(4).unwrap(),
+        );
         let user_context_weak = Arc::downgrade(&user_context);
         let this = Arc::new_cyclic(|this| Self {
             this: Weak::clone(this),
@@ -69,7 +77,8 @@ impl MailUserContext {
             action_queue,
             event_loop: EventLoop::new(),
             prefetch: OnceLock::new(),
-            default_queue_executor: queue_executor,
+            default_queue_executor: queue_executor.into_auto_executor(),
+            send_queue_executors,
         });
 
         this.action_queue
@@ -123,18 +132,15 @@ impl MailUserContext {
         self.user_context.session()
     }
 
-    pub async fn execute_all_actions(&self) -> QueuedResult<usize> {
-        self.default_queue_executor.execute_all().await
-    }
-
     /// Get the action queue instance.
     pub fn action_queue(&self) -> &Queue {
         &self.action_queue
     }
 
-    /// Get the default queue executor.
-    pub fn default_queue_executor(&self) -> &QueueExecutor {
-        &self.default_queue_executor
+    /// Terminate all action queue executors.
+    pub fn terminate_queue_executors(&self) {
+        self.default_queue_executor.terminate();
+        self.send_queue_executors.terminate();
     }
 
     /// Get the API service.