+++ mail/mail-common/src/user_context/action_queue.rs
@@ -16,17 +16,6 @@ impl MailUserContext {
     ) -> MailContextResult<QueuedActionOutput<T>> {
         Ok(self.action_queue.queue_action(action).await?)
     }
-
-    /// Execute exactly one pending action in the queue.
-    pub async fn execute_pending_action(&self) -> MailContextResult<()> {
-        self.default_queue_executor.execute_one().await?;
-        Ok(())
-    }
-
-    /// Execute all pending actions in the queue.
-    pub async fn execute_pending_actions(&self) -> MailContextResult<usize> {
-        Ok(self.default_queue_executor.execute_all().await?)
-    }
 }
 
 pub(super) async fn new_action_queue(stash: Stash) -> proton_action_queue::queue::Result<Queue> {