+++ mail/mail-common/tests/actions_read_unread.rs
@@ -9,7 +9,7 @@ use proton_mail_common::models::Conversation;
 use proton_mail_common::models::ConversationCounters;
 use proton_mail_common::Mailbox;
 use proton_mail_test_utils::init::Params;
-use proton_mail_test_utils::test_context::MailTestContext;
+use proton_mail_test_utils::test_context::{MailTestContext, MailUserContextTestExtension};
 use stash::orm::Model;
 use stash::params;
 use std::sync::LazyLock;
@@ -163,11 +163,7 @@ async fn mark_conversation_read(conversations: &[TestItem], expected_read: usize
     )
     .await
     .unwrap();
-    user_ctx
-        .default_queue_executor()
-        .execute_one()
-        .await
-        .unwrap();
+    user_ctx.execute_single_action().await.unwrap();
 
     // Validation
     let conversations = Conversation::find("WHERE num_unread = ?", params![0], &tether)
@@ -244,11 +240,7 @@ async fn mark_conversation_unread(conversations: &[TestItem], expected_read: usi
     )
     .await
     .unwrap();
-    user_ctx
-        .default_queue_executor()
-        .execute_one()
-        .await
-        .unwrap();
+    user_ctx.execute_single_action().await.unwrap();
 
     // Validation
     let conversations = Conversation::find("WHERE num_unread = ?", params![0], &tether)